requirements.txt
fastapi==0.109.0
sqlalchemy==2.0.25
passlib==1.7.4
python-jose==3.3.0
pydantic==2.5.3
bcrypt==4.1.2
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
uvicorn==0.27.0
email-validator==2.2.0
psycopg2-binary-2.9.10
Pillow==10.2.0  # For certificate generation
reportlab==4.1.0  # For PDF generation